+++ core/src/python_wrapper/py_score.cpp
@@ -87,6 +87,10 @@ void ScoreClass(py::module &m) {
     cls.def("getPartName", &Score::getPartName, py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
     cls.def("selectNotes", &Score::selectNotes, py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
     cls.def("countNotes", &Score::countNotes, py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
+    cls.def("setRepeat", &Score::setRepeat, 
+        py::arg("measureStart"),
+        py::arg("measureEnd") = -1);
+
     cls.def("findPattern", &Score::findPattern, py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
     cls.def("instrumentFragmentation", &Score::instrumentFragmentation, py::call_guard<py::scoped_ostream_redirect, py::scoped_estream_redirect>());
     cls.def("getChords", [](Score& score, nlohmann::json config)